lxd/fsmonitor: Move events out of drivers package.
